Web scraping project
The aim of this project is to compare coverage of international conflicts in Israel in Ukraine. Sentences containing the names Biden, Netanyahu(for Israel) and Putin (for Ukraine) are contrasted with eachother. In addition, descriptive statistics were collected from the scraped AP news articles, as well as a Intertopic Model from the LDAvis package.